Creation of the class VariationalOptimalControlProgram
