+++ bioptim/examples/discrete_mechanics_for_optimal_control/pendulum.py
@@ -4,22 +4,17 @@ variational integrator. Moreover, the model has been freed on the z-axis, it is 
 """
 from bioptim import (
     Bounds,
-    DynamicsList,
     InitialGuess,
     InterpolationType,
     Objective,
     ObjectiveFcn,
-    ParameterList,
     Solver,
 )
 import matplotlib.pyplot as plt
 import numpy as np
-import pickle
-
-from variational_integrator import *
-from save_results import save_results
 
 from biorbd_model_holonomic import BiorbdModelCustomHolonomic
+from variational_optimal_control_program import VariationalOptimalControlProgram
 
 
 def prepare_ocp(
@@ -27,7 +22,7 @@ def prepare_ocp(
     final_time: float,
     n_shooting: int,
     use_sx: bool = True,
-) -> OptimalControlProgram:
+) -> VariationalOptimalControlProgram:
     """
     The initialization of an ocp
 
@@ -70,52 +65,27 @@ def prepare_ocp(
     # Initial guess
     u_init = InitialGuess([0] * n_q)
 
-    # Declare parameters for the initial and final velocities
-    parameters = ParameterList()
-    # Give the parameter some min and max bounds
+    # Give the initial and final some min and max bounds
     qdot0_bounds = Bounds([0.0, 0.0], [0.0, 0.0], interpolation=InterpolationType.CONSTANT)
     qdotN_bounds = Bounds([0.0, 0.0], [0.0, 0.0], interpolation=InterpolationType.CONSTANT)
     # And an initial guess
     qdot0_init = InitialGuess([0] * n_q)
     qdotN_init = InitialGuess([0] * n_q)
 
-    parameters.add(
-        "qdot0",  # The name of the parameter
-        function=qdot_function,  # The function that modifies the biorbd model
-        initial_guess=qdot0_init,  # The initial guess
-        bounds=qdot0_bounds,  # The bounds
-        size=n_q,  # The number of elements this particular parameter vector has
-    )
-    parameters.add(
-        "qdotN",  # The name of the parameter
-        function=qdot_function,  # The function that modifies the biorbd model
-        initial_guess=qdotN_init,  # The initial guess
-        bounds=qdotN_bounds,  # The bounds
-        size=n_q,  # The number of elements this particular parameter vector has
-    )
-
-    # Dynamics
-    dynamics = DynamicsList()
-    expand = True
-    dynamics.add(custom_configure_unconstrained, bio_model=bio_model, expand=expand)
-
-    multinode_constraints = variational_continuity(n_shooting, n_q)
-
-    return OptimalControlProgram(
+    return VariationalOptimalControlProgram(
         bio_model,
-        dynamics,
         n_shooting,
         final_time,
-        x_init=x_init,
+        q_init=x_init,
         u_init=u_init,
-        x_bounds=x_bounds,
+        q_bounds=x_bounds,
         u_bounds=u_bounds,
+        qdot0_init=qdot0_init,
+        qdot0_bounds=qdot0_bounds,
+        qdotN_init=qdotN_init,
+        qdotN_bounds=qdotN_bounds,
         objective_functions=objective_functions,
         use_sx=use_sx,
-        assume_phase_dynamics=True,
-        skip_continuity=True,
-        parameters=parameters,
-        multinode_constraints=multinode_constraints,
     )
